requirements.txt
# 核心依赖
pandas >= 1.5.3
numpy >= 1.24.3
scikit-learn >= 1.2.2
requests >= 2.31.0
beautifulsoup4 >= 4.12.2
python-dotenv >= 1.0.0
loguru >= 0.7.0
pydantic >= 1.10.7
langchain >= 0.1.16
crewai >= 0.1.12
akshare >= 1.11.50
pandasai >= 2.0.0
google-search-results >= 2.4.2
igraph >= 0.10.4
langgraph >= 0.0.10
xgboost >= 1.7.5
pyanp >= 0.3.1
langchain-core >= 0.3.74
langchain-community >= 0.0.38
langsmith >= 0.4.14

# 开发依赖
pytest >= 7.3.1
flake8 >= 6.0.0
black >= 23.3.0
mypy >= 1.2.0